backend/requirements.txt
flask
flask-cors
pymongo
dnspython
python-dotenv
openpyxl
flask-login